pyproject.toml
[project]
name = "elite-voice-agent"
version = "0.1.0"
description = "Elite Portfolio Management — Inbound voice agent powered by LiveKit"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[openai,silero,deepgram,turn-detector]~=1.4",
    "livekit-api>=1.0",
    "telnyx-livekit-plugin @ git+https://github.com/team-telnyx/telnyx-livekit-plugin.git#subdirectory=telnyx-livekit-plugin",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
